+++ lib/generate/json2dart/generators/response_model_generator.dart
@@ -76,6 +76,69 @@ ${_generateNestedClasses(map, suffix, apiClassName)}''';
     return classContent;
   }
 
+  /// Generates a complete response model file
+  ///
+  /// [className] - Main class name
+  /// [data] - JSON data structure
+  /// Returns the complete file content
+  String generateExtraModel(
+    String className,
+    Map<String, dynamic> data,
+  ) {
+    _listClassName.clear();
+
+    String imports = '''import 'dart:convert';
+
+import 'package:core/core.dart';
+
+''';
+
+    final classContent = generateExtraClass(className, 'Extra', '', data, true);
+    return imports + classContent;
+  }
+
+  String generateExtraClass(
+    String suffix,
+    String name,
+    String parent,
+    Map<String, dynamic>? map, [
+    bool isRoot = false,
+  ]) {
+    if (map == null) return '';
+
+    String apiClassName = ModelClassNameHelper.getClassName(
+      _listClassName,
+      suffix,
+      name,
+      isRoot,
+      true,
+      parent,
+    );
+
+    apiClassName = '${apiClassName.replaceAll('Extra', '')}Extra';
+
+    final classContent = '''class $apiClassName extends Equatable {
+  ${generateConstructor(apiClassName, map)}
+
+  ${_generateFromExtraMap(apiClassName, map, suffix, apiClassName)}
+
+  factory $apiClassName.fromJson(String source) =>
+      $apiClassName.fromMap(json.decode(source));
+
+  ${_generateTypeExtraData(map, suffix, apiClassName)}
+
+  ${_generateToMap(map)}
+
+  String toJson() => json.encode(toMap());
+
+  ${generateProps(map)}
+}
+
+${_generateNestedExtraClasses(map, suffix, apiClassName)}''';
+
+    return classContent;
+  }
+
   /// Generates fromMap factory constructor
   String _generateFromMap(
     String className,
@@ -92,6 +155,22 @@ ${_generateNestedClasses(map, suffix, apiClassName)}''';
   }''';
   }
 
+  /// Generates fromMap factory constructor
+  String _generateFromExtraMap(
+    String className,
+    Map<String, dynamic> map,
+    String suffix,
+    String parent,
+  ) {
+    final variables = map.keys;
+
+    return '''factory $className.fromMap(Map<String, dynamic> map,) {
+    return ${variables.isEmpty ? 'const' : ''} $className(
+      ${variables.map((e) => "${e.toString().camelCase}: ${_getVariableFromExtraMap(e, map[e], suffix, parent)}").join(',      \n')}${variables.isNotEmpty ? ',' : ''}
+    );
+  }''';
+  }
+
   /// Generates type definitions for response model
   String _generateTypeData(
     Map<String, dynamic> map,
@@ -103,6 +182,24 @@ ${_generateNestedClasses(map, suffix, apiClassName)}''';
     String result = '';
     for (final variable in variables) {
       final type = getTypeVariable(
+          variable, map[variable], suffix, _listClassName, parent);
+      final isNullable = type != 'dynamic' ? '?' : '';
+      result += '  final $type$isNullable ${variable.toString().camelCase};\n';
+    }
+
+    return result;
+  }
+
+  String _generateTypeExtraData(
+    Map<String, dynamic> map,
+    String suffix,
+    String parent,
+  ) {
+    final variables = map.keys;
+
+    String result = '';
+    for (final variable in variables) {
+      final type = getTypeExtraVariable(
           variable, map[variable], suffix, _listClassName, parent);
       final isNullable = type != 'dynamic' ? '?' : '';
       result += '  final $type$isNullable ${variable.toString().camelCase};\n';
@@ -177,6 +274,65 @@ ${_generateNestedClasses(map, suffix, apiClassName)}''';
     return variable;
   }
 
+  /// Gets the appropriate deserialization format for response variables
+  String _getVariableFromExtraMap(
+      String key, dynamic value, String suffix, String parent) {
+    final variable = "map['$key']";
+
+    if (value is int) {
+      return "int.tryParse($variable?.toString() ?? '')";
+    }
+
+    if (value is double) {
+      return "double.tryParse($variable?.toString() ?? '')";
+    }
+
+    if (value is bool) {
+      return variable;
+    }
+
+    if (value is Map) {
+      String className = ModelClassNameHelper.getClassName(
+        _listClassName,
+        suffix,
+        key.pascalCase,
+        false,
+        false,
+        parent,
+      );
+      className = '${className.replaceAll('Extra', '')}Extra';
+
+      final data = '$className.fromMap($variable,)';
+      return '$variable == null ? null : $data';
+    }
+
+    if (value is List && value.isNotEmpty) {
+      if (value.first is Map) {
+        String className = ModelClassNameHelper.getClassName(
+          _listClassName,
+          suffix,
+          key.pascalCase,
+          false,
+          false,
+          parent,
+        );
+        className = '${className.replaceAll('Extra', '')}Extra';
+
+        final data =
+            'List.from(($variable as List).where((element) => element != null).map((e) => $className.fromMap(e)),)';
+        return '$variable is List ? $data : null';
+      } else {
+        return _handlePrimitiveList(variable, value.first);
+      }
+    }
+
+    if (value is String && _isDateTime(value)) {
+      return "DateTime.tryParse($variable ?? '')";
+    }
+
+    return variable;
+  }
+
   /// Handles primitive list deserialization
   String _handlePrimitiveList(String variable, dynamic firstElement) {
     if (firstElement is int) {
@@ -228,6 +384,37 @@ ${_generateNestedClasses(map, suffix, apiClassName)}''';
         ));
       } else if (value is List && value.isNotEmpty && value.first is Map) {
         nestedClasses.add(generateClass(
+          suffix,
+          key.toString().pascalCase,
+          parent,
+          value.first,
+          false,
+        ));
+      }
+    }
+
+    return nestedClasses.join('\n');
+  }
+
+  /// Generates nested classes for complex objects
+  String _generateNestedExtraClasses(
+      Map<String, dynamic> map, String suffix, String parent) {
+    final nestedClasses = <String>[];
+
+    for (final entry in map.entries) {
+      final key = entry.key;
+      final value = entry.value;
+
+      if (value is Map<String, dynamic>) {
+        nestedClasses.add(generateExtraClass(
+          suffix,
+          key.toString().pascalCase,
+          parent,
+          value,
+          false,
+        ));
+      } else if (value is List && value.isNotEmpty && value.first is Map) {
+        nestedClasses.add(generateExtraClass(
           suffix,
           key.toString().pascalCase,
           parent,